# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Online Library Documentation",
    "version": "0.0.1",
    "description": "Online Library"
  },
  "paths": {
    "/users": {
      "get": {
        "tags": [
          "Users"
        ],
        "description": "Get all users",
        "operationId": "getAllUsers",
        "parameters": [
          {
            "name": "perPage",
            "in": "query",
            "description": "How many userss will be showed in one page",
            "required": false,
            "default": 20,
            "type": "number"
          },
          {
            "name": "page",
            "in": "query",
            "description": "page number",
            "required": false,
            "default": 1,
            "type": "number"
          },
          {
            "name": "name",
            "in": "query",
            "description": "user name",
            "required": false,
            "default": "Maria",
            "type": "string"
          },
          {
            "name": "email",
            "in": "query",
            "description": "user email",
            "required": false,
            "default": "olga@gmail.com",
            "type": "string"
          },
          {
            "name": "role",
            "in": "query",
            "description": "customer role",
            "required": false,
            "default": "user",
            "type": "array",
            "enum": ["user", "admin"],
            "collectionFormat": "multi"
          }],
        "responses": {
          "200": {
            "description": "Success response",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/singleUser"
              }
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Email is not valid"
                }
              }
            }
          }
        }
      }
    },
    "/users/{userId}": {
      "get": {
        "tags": [
          "User's info"
        ],
        "description": "Get info about user by id",
        "operationId": "get one User",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": "true",
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": "1"
            }
          },
          {
            "name": "authorization token",
            "in": "authorization header",
            "required": "true",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success response",
            "schema": {
              "type": "object",
              "$ref": "#/definitions/singleUser"
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Email is not valid"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized user",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Unauthorized user"
                }
              }
            }
          }
        }
      }
    }
  },
  "definitions": {
    "singleUser": {
      "type": "object",
      "properties": {
        "_id": {
          "description": "customer ID",
          "type": "string",
          "example": "228"
        },
        "name": {
          "description": "customer name",
          "type": "string",
          "example": "Olga"
        },
        "email": {
          "description": "customer email",
          "type": "string",
          "example": "admin@inoxoft.com"
        },
        "role": {
          "description": "role of user",
          "type": "string",
          "example": "admin"
        },
        "book": {
          "description": "user's books",
          "type": "array",
          "items": {
            "description": "books ID",
            "type": "string",
            "example": "228"
          }
        },
        "createdAt": {
          "description": "Created at",
          "type": "string",
          "example": "2021-09-14T07:56:30.115Z"
        },
        "updatedAt": {
          "description": "Updeted at",
          "type": "string",
          "example": "2021-09-14T07:56:30.115Z"
        },
        "avatar": {
          "description": "avatar of user",
          "type": "url",
          "example": "http://dfdfdg"
        }
      }
    },
    "errorSchema": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "example": "Email is not valid"
        },
        "data": {
          "description": "customer email",
          "type": "string",
          "example": "admin@inoxoft.com"
        }
      }
    }
  }
}